get rid of MonadPlus instance for RandT.
It actually does not make a whole lot of sense, because the semantics
of m1 `mplus` m2, where m1 and m2 are of a type like RandT Maybe a, is
to run m1 and m2 with the *same* seed.  This does not seem very
useful (and can easily lead to bugs).  Much better is to use  MaybeT Rand.